Create csample.pl
This script creates simple random samples or systematic random samples from concordances created with the Open Corpus Workbench, with the option of specifying the selection criteria or a desired sample size. It has options for including or excluding the standard  CWB header (and will insert information about sample size and sample method into the header).